Getting rid of all longs in the Prime library; let Python manage the numeric types. besides, forcing longs makes the calculation take that much longer.
also adding P95. it's slow atm (11 seconds on pypy 1.8)

+++ lib/python/ProjectEulerPrime.py
@@ -15,8 +15,8 @@ class ProjectEulerPrime(Prime):
         Prime.__init__(self)
     
     def isPrime(self, n):
-        n = long(n)
-        if n < 2L:
+        n = int(n)
+        if n < 2:
         	return False
         elif not Prime.factor(self, n):
             return True
